pyproject.toml
[project]
name = "cryptoforge"
version = "0.1.0"
description = "Modular, multi-strategy, event-driven cryptocurrency trading bot"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.0",
    "redis>=5.0",
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "pandas>=2.1",
    "pandas-ta>=0.3.14b1",
    "numpy>=1.26",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "orjson>=3.9",
    "python-telegram-bot>=20.7",
    "prometheus-client>=0.19",
    "httpx>=0.26",
    "tenacity>=8.2",
    "structlog>=24.1",
    "rich>=13.7",
    "click>=8.1",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
]
ml = [
    "xgboost>=2.0",
    "scikit-learn>=1.4",
    "hmmlearn>=0.3",
    "vectorbt>=0.26",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
